fix(promotion) : promotion discountRate 변경

+++ order/src/main/java/com/ohgiraffers/order/core/api/command/response/PromotionResponse.java
@@ -11,7 +11,7 @@ public record PromotionResponse (
         String productName,
         Integer originalPrice,
         Integer salePrice,
-        Double discountRate,
+        Integer discountRate,
         Integer totalQuantity,
         Integer soldQuantity,
         @JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss")
